Create beta version in main directory
This is because it seems if the beta version is put in the Working directory
we can't install it. However if it is in this, the HomebrewFormula directory,
the issue is that it conflicts with the xmds2.rb install. So the beta version should
be removed from this directory when not in use.